RAG + agents/requirements.txt
# FastAPI
fastapi
uvicorn[standard]
python-multipart
python-dotenv

# Database
sqlalchemy
psycopg2-binary
alembic

# Authentication
python-jose[cryptography]
passlib[bcrypt]
argon2-cffi

# Agents
redis
openai
langchain
langchain-openai
langchain-community
langchain-huggingface
faiss-cpu
langchain-groq
tiktoken

# Voice (inference only)
elevenlabs
openai-whisper
gtts
pydub
librosa

# Utilities
numpy
requests
email-validator

# Missing System Dependencies from teammates
pandas
transformers
sentence-transformers

# Testing
pytest
httpx

# NOTE: ffmpeg must also be installed separately on your system (not on venv)
# Windows: winget install ffmpeg
# Verify: ffmpeg -version